ver0008 authentication to column and task routes

+++ routes/column.js
@@ -5,19 +5,26 @@ const router = express.Router();
 
 const { Column, Task, User } = require('../db/models/index');
 
+const authMiddleware = require('../middleware/auth-middleware');
+
 // Gets all columns
-router.get('/', (req, res) => {
-    Column.find({}).then((columns) => {
+router.get('/', authMiddleware.authenticate, (req, res) => {
+    Column.find({
+        _userId: req.user_id
+    }).then((columns) => {
         res.send(columns);
+    }).catch((e) => {
+        res.send(e);
     });
 })
 
 // Creates a new column
-router.post('/', (req, res) => {
+router.post('/', authMiddleware.authenticate, (req, res) => {
     let title = req.body.title;
 
     let newColumn = new Column({
-        title
+        title,
+        _userId: req.user_id
     });
 
     newColumn.save().then((createdColumn) => {
@@ -26,9 +33,9 @@ router.post('/', (req, res) => {
 });
 
 // Updates a column
-router.put('/:id', (req, res) => {
+router.put('/:id', authMiddleware.authenticate, (req, res) => {
     Column.findOneAndUpdate(
-        {_id: req.params.id},
+        {_id: req.params.id, _userId: req.user_id},
         {$set: req.body}
     ).then(() => {
         res.sendStatus(200);
@@ -36,11 +43,17 @@ router.put('/:id', (req, res) => {
 });
 
 // Deletes a column
-router.delete('/:id', (req, res) => {
-    Column.findOneAndRemove({_id: req.params.id}).then((removedColumn) => {
+router.delete('/:id', authMiddleware.authenticate, (req, res) => {
+    Column.findOneAndRemove({_id: req.params.id, _userId: req.user_id}).then((removedColumn) => {
         res.send(removedColumn);
+        deleteAllTasksFromColumn(removedColumn._id);
     });
 });
+
+// Helper method to delete all tasks in a column
+let deleteAllTasksFromColumn = (_columnId) => {
+    Task.deleteMany({_columnId});
+}
 
 
 module.exports = router;
